refactor: demote services/executor.ts + document World B elimination path
Extract SafePgCallFn to lib/safe-pg-call.ts (fixes cross-layer import from
lib/audit-events → services/executor). Slim executor.ts from 99 to 43 lines
by removing verbose JSDoc and section separators. Add TODO markers on
OmniSession and IExecutor pointing to World A replacements.

OmniSession references now confined to executor.ts + 3 direct consumers
(omni-bridge, claude-code executor, claude-sdk executor) — no leakage
outside src/services/.
