+++ knip.json
@@ -5,6 +5,7 @@
     "src/tui/render.tsx",
     "src/tui/tmux.ts",
     "src/lib/executor-types.ts",
+    "src/lib/safe-pg-call.ts",
     "src/lib/executor-registry.ts",
     "src/lib/assignment-registry.ts",
     "src/lib/providers/*.ts",
